Introducing one more level of abstraction to be able to provide a more efficient rescheduling
